Created an isolated module of operation_intent.
